Move get_category_options to utils
